requirements.txt
opencv-python
numpy
imread-from-url
cap_from_youtube
onnxruntime-gpu